requirements.txt
streamlit==1.46.1
pandas==2.3.0
pillow==11.3.0
google-generativeai==0.8.5
python-dotenv==1.1.1
plotly==6.2.0